refactor: :recycle: remove unnecessary hook
