+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/background.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/Bucket.tsx","./src/components/CustomSwitch.tsx","./src/components/TabBuckets.tsx","./src/components/TabButton.tsx","./src/components/ui/accordion.tsx","./src/components/ui/badge.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/separator.tsx","./src/components/ui/switch.tsx","./src/components/ui/tabs.tsx","./src/hooks/useLastSeen.hook.ts","./src/lib/utils.ts","./src/store/useBuckets.store.ts","./src/store/useLastSeen.store.ts","./src/types/index.ts"],"version":"5.6.2"}
+{"root":["./src/App.tsx","./src/background.ts","./src/main.tsx","./src/vite-env.d.ts","./src/components/Bucket.tsx","./src/components/CustomSwitch.tsx","./src/components/TabBuckets.tsx","./src/components/TabButton.tsx","./src/components/ui/accordion.tsx","./src/components/ui/badge.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/separator.tsx","./src/components/ui/switch.tsx","./src/components/ui/tabs.tsx","./src/lib/utils.ts","./src/store/useBuckets.store.ts","./src/store/useLastSeen.store.ts","./src/types/index.ts"],"version":"5.6.2"}
